\name{reglogit-internal}
%% Part of the sensible export list but (currently?) documented as
%% internal (or waiting for documentation to be written):
%% <LIST>
\alias{draw.beta}
\alias{draw.nu}
\alias{draw.z}
\alias{draw.omega}
\alias{draw.lambda}
\alias{my.rinvgauss}
\alias{calc.lpost}
\alias{preprocess}
\alias{z.dRUM}
\alias{beta.dRUM}
\alias{gibbs.dRUM}
\alias{calc.Cs}
\alias{calc.mlpost}
\alias{mpreprocess}
\alias{rmultnorm}
\title{Internal reglogit Functions}
\description{
  Internal \pkg{reglogit} functions
}
\details{
  These functions are primarily provided to facilitate the
  demos, and to provide for a cleaner exposition.  We encourage
  users to inspect their contents in order to help develop 
  similar functions to suit their particular needs
}
\keyword{ internal }
